+++ CHANGELOG.md
@@ -7,6 +7,13 @@
 - add warp from teacher's lounge to cpu wing
 - add soul upgrades
 - add purple pen rewards from Annie's Boobs
+- add purple pen upgrades:
+  - Soul Catcher
+  - Quick Focus
+  - Dream Wielder
+  - Deep Focus
+  - Shaman Stone
+  - Spell Twister
 
 ## [0.0.10] - 2024-01-16